update example and tests to also include parallel use case
