--- sdk/samples/sequential/sequential.py
@@ -1,53 +1,0 @@
-#!/usr/bin/env python3
-
-# Copyright 2019-2020 kubeflow.org
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from kfp_tekton.compiler import TektonCompiler
-from kfp import dsl
-
-
-def gcs_download_op(url):
-    return dsl.ContainerOp(
-        name='GCS - Download',
-        image='google/cloud-sdk:216.0.0',
-        command=['sh', '-c'],
-        arguments=['gsutil cat $0 | tee $1', url, '/tmp/results.txt'],
-    )
-
-
-def echo_op(text):
-    return dsl.ContainerOp(
-        name='echo',
-        image='library/bash:4.4.23',
-        command=['sh', '-c'],
-        arguments=['echo "$0"', text]
-    )
-
-
-@dsl.pipeline(
-    name='Sequential pipeline',
-    description='A pipeline with two sequential steps.'
-)
-def sequential_pipeline(url='gs://ml-pipeline-playground/shakespeare1.txt', path='/tmp/results.txt'):
-    """A pipeline with two sequential steps."""
-
-    download_task = gcs_download_op(url)
-    echo_task = echo_op(path)
-
-    echo_task.after(download_task)
-
-
-if __name__ == '__main__':
-    TektonCompiler().compile(sequential_pipeline, 'sequential.yaml')
